+++ src/main/java/miyucomics/hexical/features/telepathy/OpGetKeybindSelf.kt
@@ -7,7 +7,7 @@ import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.iota.NullIota
 import net.minecraft.server.network.ServerPlayerEntity
 
-class OpGetKeybind(private val key: String) : ConstMediaAction {
+class OpGetKeybindSelf(private val key: String) : ConstMediaAction {
 	override val argc = 0
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
 		val caster = env.castingEntity